+++ agent/backend/pktvisor/pktvisor.go
@@ -5,13 +5,9 @@
 package pktvisor
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"os/exec"
 	"strings"
@@ -22,13 +18,9 @@ import (
 	"github.com/go-co-op/gocron"
 	"github.com/ns1labs/orb/agent/backend"
 	"github.com/ns1labs/orb/agent/config"
-	"github.com/ns1labs/orb/agent/otel/otlpmqttexporter"
-	"github.com/ns1labs/orb/agent/otel/pktvisorreceiver"
 	"github.com/ns1labs/orb/agent/policies"
-	"github.com/ns1labs/orb/fleet"
 	"go.opentelemetry.io/collector/component"
 	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
 )
 
 var _ backend.Backend = (*pktvisorBackend)(nil)
@@ -39,10 +31,18 @@ const (
 	ReadinessTimeout    = 10
 	ApplyPolicyTimeout  = 10
 	RemovePolicyTimeout = 15
-	VersionTimeout      = 5
+	VersionTimeout      = 2 // also used for readiness check
 	ScrapeTimeout       = 5
 	TapsTimeout         = 5
 )
+
+// AppInfo represents server application information
+type AppInfo struct {
+	App struct {
+		Version   string  `json:"version"`
+		UpTimeMin float64 `json:"up_time_min"`
+	} `json:"app"`
+}
 
 type pktvisorBackend struct {
 	logger          *zap.Logger
@@ -85,161 +85,29 @@ func (p *pktvisorBackend) SetCommsClient(agentID string, client mqtt.Client, bas
 	p.otlpMetricsTopic = fmt.Sprintf("%s/m/%c", otelMetricsTopic, agentID[0])
 }
 
-func (p *pktvisorBackend) GetState() (backend.BackendState, string, error) {
-	_, err := p.checkAlive()
-	if err != nil {
-		return backend.Unknown, "", err
+func (p *pktvisorBackend) GetRunningStatus() (backend.RunningStatus, string, error) {
+	// first check process status
+	runningStatus, errMsg, err := p.getProcRunningStatus()
+	// if it's not running, we're done
+	if runningStatus != backend.Running {
+		return runningStatus, errMsg, err
 	}
-	return backend.Running, "", nil
-}
-
-// AppMetrics represents server application information
-type AppMetrics struct {
-	App struct {
-		Version   string  `json:"version"`
-		UpTimeMin float64 `json:"up_time_min"`
-	} `json:"app"`
-}
-
-// note this needs to be stateless because it is called for multiple go routines
-func (p *pktvisorBackend) request(url string, payload interface{}, method string, body io.Reader, contentType string, timeout int32) error {
-	client := http.Client{
-		Timeout: time.Second * time.Duration(timeout),
+	// if it's running, check REST API availability too
+	_, aiErr := p.getAppInfo()
+	if aiErr != nil {
+		// process is running, but REST API is not accessible
+		return backend.BackendError, "process running, REST API unavailable", aiErr
 	}
-
-	alive, err := p.checkAlive()
-	if !alive {
-		return err
-	}
-
-	URL := fmt.Sprintf("%s://%s:%s/api/v1/%s", p.adminAPIProtocol, p.adminAPIHost, p.adminAPIPort, url)
-
-	req, err := http.NewRequest(method, URL, body)
-	if err != nil {
-		p.logger.Error("received error from payload", zap.Error(err))
-		return err
-	}
-
-	req.Header.Add("Content-Type", contentType)
-	res, getErr := client.Do(req)
-
-	if getErr != nil {
-		p.logger.Error("received error from payload", zap.Error(getErr))
-		return getErr
-	}
-
-	if (res.StatusCode < 200) || (res.StatusCode > 299) {
-		body, err := ioutil.ReadAll(res.Body)
-		if err != nil {
-			return errors.New(fmt.Sprintf("non 2xx HTTP error code from pktvisord, no or invalid body: %d", res.StatusCode))
-		}
-		if len(body) == 0 {
-			return errors.New(fmt.Sprintf("%d empty body", res.StatusCode))
-		} else if body[0] == '{' {
-			var jsonBody map[string]interface{}
-			err := json.Unmarshal(body, &jsonBody)
-			if err == nil {
-				if errMsg, ok := jsonBody["error"]; ok {
-					return errors.New(fmt.Sprintf("%d %s", res.StatusCode, errMsg))
-				}
-			}
-		}
-	}
-
-	if res.Body != nil {
-		err = json.NewDecoder(res.Body).Decode(&payload)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (p *pktvisorBackend) checkAlive() (bool, error) {
-	status := p.proc.Status()
-
-	if status.Error != nil {
-		p.logger.Error("pktvisor process error", zap.Error(status.Error))
-		return false, status.Error
-	}
-
-	if status.Complete {
-		err := p.proc.Stop()
-		if err != nil {
-			p.logger.Error("proc.Stop error", zap.Error(err))
-		}
-		return false, errors.New("pktvisor process ended")
-	}
-
-	if status.StopTs > 0 {
-		p.logger.Info("pktvisor process stopped")
-		return false, errors.New("pktvisor process ended")
-	}
-
-	return true, nil
-}
-
-func (p *pktvisorBackend) ApplyPolicy(data policies.PolicyData, updatePolicy bool) error {
-
-	if updatePolicy {
-		// To update a policy it's necessary first remove it and then apply a new version
-		err := p.RemovePolicy(data)
-		if err != nil {
-			p.logger.Warn("policy failed to remove", zap.String("policy_id", data.ID), zap.String("policy_name", data.Name), zap.Error(err))
-		}
-	}
-
-	p.logger.Debug("pktvisor policy apply", zap.String("policy_id", data.ID), zap.Any("data", data.Data))
-
-	fullPolicy := map[string]interface{}{
-		"version": "1.0",
-		"visor": map[string]interface{}{
-			"policies": map[string]interface{}{
-				data.Name: data.Data,
-			},
-		},
-	}
-
-	policyYaml, err := yaml.Marshal(fullPolicy)
-	if err != nil {
-		p.logger.Warn("yaml policy marshal failure", zap.String("policy_id", data.ID), zap.Any("policy", fullPolicy))
-		return err
-	}
-
-	var resp map[string]interface{}
-	err = p.request("policies", &resp, http.MethodPost, bytes.NewBuffer(policyYaml), "application/x-yaml", ApplyPolicyTimeout)
-	if err != nil {
-		p.logger.Warn("yaml policy application failure", zap.String("policy_id", data.ID), zap.ByteString("policy", policyYaml))
-		return err
-	}
-
-	return nil
-
-}
-
-func (p *pktvisorBackend) RemovePolicy(data policies.PolicyData) error {
-	var resp interface{}
-	err := p.request(fmt.Sprintf("policies/%s", data.Name), &resp, http.MethodDelete, http.NoBody, "application/json", RemovePolicyTimeout)
-	if err != nil {
-		p.logger.Error("received error", zap.Error(err))
-		return err
-	}
-	return nil
+	return runningStatus, "", nil
 }
 
 func (p *pktvisorBackend) Version() (string, error) {
-	var appMetrics AppMetrics
-	err := p.request("metrics/app", &appMetrics, http.MethodGet, http.NoBody, "application/json", VersionTimeout)
+	appInfo, err := p.getAppInfo()
 	if err != nil {
 		return "", err
 	}
-	p.pktvisorVersion = appMetrics.App.Version
-	return appMetrics.App.Version, nil
-}
-
-func (p *pktvisorBackend) Write(payload []byte) (n int, err error) {
-	p.logger.Info("pktvisord", zap.ByteString("log", payload))
-	return len(payload), nil
+	p.pktvisorVersion = appInfo.App.Version
+	return appInfo.App.Version, nil
 }
 
 func (p *pktvisorBackend) Start(ctx context.Context, cancelFunc context.CancelFunc) error {
@@ -325,7 +193,7 @@ func (p *pktvisorBackend) Start(ctx context.Context, cancelFunc context.CancelFu
 
 	var readinessError error
 	for backoff := 0; backoff < ReadinessBackoff; backoff++ {
-		var appMetrics AppMetrics
+		var appMetrics AppInfo
 		readinessError = p.request("metrics/app", &appMetrics, http.MethodGet, http.NoBody, "application/json", ReadinessTimeout)
 		if readinessError == nil {
 			p.logger.Info("pktvisor readiness ok, got version ", zap.String("pktvisor_version", appMetrics.App.Version))
@@ -357,131 +225,6 @@ func (p *pktvisorBackend) Start(ctx context.Context, cancelFunc context.CancelFu
 	}
 
 	return nil
-}
-
-func (p *pktvisorBackend) scrapeDefault() error {
-	// scrape all policy json output with one call every minute.
-	// TODO support policies with custom bucket times
-	job, err := p.scraper.Every(1).Minute().WaitForSchedule().Do(func() {
-		metrics, err := p.scrapeMetrics(1)
-		if err != nil {
-			p.logger.Error("scrape failed", zap.Error(err))
-			return
-		}
-		if len(metrics) == 0 {
-			p.logger.Warn("scrape: no policies found, skipping")
-			return
-		}
-
-		var batchPayload []fleet.AgentMetricsRPCPayload
-		totalSize := 0
-		for pName, pMetrics := range metrics {
-			data, err := p.policyRepo.GetByName(pName)
-			if err != nil {
-				p.logger.Warn("skipping pktvisor policy not managed by orb", zap.String("policy", pName), zap.String("error_message", err.Error()))
-				continue
-			}
-			payloadData, err := json.Marshal(pMetrics)
-			if err != nil {
-				p.logger.Error("error marshalling scraped metric json", zap.String("policy", pName), zap.Error(err))
-				continue
-			}
-			metricPayload := fleet.AgentMetricsRPCPayload{
-				PolicyID:   data.ID,
-				PolicyName: data.Name,
-				Datasets:   data.GetDatasetIDs(),
-				Format:     "json",
-				BEVersion:  p.pktvisorVersion,
-				Data:       payloadData,
-			}
-			batchPayload = append(batchPayload, metricPayload)
-			totalSize += len(payloadData)
-			p.logger.Info("scraped metrics for policy", zap.String("policy", pName), zap.String("policy_id", data.ID), zap.Int("payload_size_b", len(payloadData)))
-		}
-
-		rpc := fleet.AgentMetricsRPC{
-			SchemaVersion: fleet.CurrentRPCSchemaVersion,
-			Func:          fleet.AgentMetricsRPCFunc,
-			Payload:       batchPayload,
-		}
-
-		// ###### DO NOT MERGE ###########
-		_, err = json.Marshal(rpc)
-		if err != nil {
-			p.logger.Error("error marshalling metric rpc payload", zap.Error(err))
-			return
-		}
-
-		p.logger.Info("FAKE METRIC SEND")
-		//if token := p.mqttClient.Publish(p.metricsTopic, 1, false, body); token.Wait() && token.Error() != nil {
-		//	p.logger.Error("error sending metrics RPC", zap.String("topic", p.metricsTopic), zap.Error(token.Error()))
-		//	return
-		//}
-		// ###### DO NO MERGE ##########
-		p.logger.Info("scraped and published metrics", zap.String("topic", p.metricsTopic), zap.Int("payload_size_b", totalSize), zap.Int("batch_count", len(batchPayload)))
-
-	})
-
-	if err != nil {
-		return err
-	}
-
-	job.SingletonMode()
-	return nil
-}
-
-func (p *pktvisorBackend) scrapeOpenTelemetry(ctx context.Context) {
-	go func() {
-		startExpCtx, cancelFunc := context.WithCancel(ctx)
-		var ok bool
-		var err error
-		for i := 1; i < 10; i++ {
-			select {
-			case <-startExpCtx.Done():
-				return
-			default:
-				if p.mqttClient != nil {
-					var errStartExp error
-					p.exporter, errStartExp = p.createOtlpMqttExporter(ctx)
-					if errStartExp != nil {
-						p.logger.Error("failed to create a exporter", zap.Error(err))
-						return
-					}
-
-					p.receiver, err = createReceiver(ctx, p.exporter, p.logger)
-					if err != nil {
-						p.logger.Error("failed to create a receiver", zap.Error(err))
-						return
-					}
-
-					err = p.exporter.Start(ctx, nil)
-					if err != nil {
-						p.logger.Error("otel mqtt exporter startup error", zap.Error(err))
-						return
-					}
-
-					err = p.receiver.Start(ctx, nil)
-					if err != nil {
-						p.logger.Error("otel receiver startup error", zap.Error(err))
-						return
-					}
-
-					ok = true
-					return
-				} else {
-					p.logger.Info("waiting until mqtt client is connected", zap.String("wait time", (time.Duration(i)*time.Second).String()))
-					time.Sleep(time.Duration(i) * time.Second)
-					continue
-				}
-			}
-		}
-		if !ok {
-			p.logger.Error("mqtt did not established a connection, stopping agent")
-			p.Stop(startExpCtx)
-		}
-		cancelFunc()
-		return
-	}()
 }
 
 func (p *pktvisorBackend) Stop(ctx context.Context) error {
@@ -535,15 +278,6 @@ func (p *pktvisorBackend) Configure(logger *zap.Logger, repo policies.PolicyRepo
 	return nil
 }
 
-func (p *pktvisorBackend) scrapeMetrics(period uint) (map[string]interface{}, error) {
-	var metrics map[string]interface{}
-	err := p.request(fmt.Sprintf("policies/__all/metrics/bucket/%d", period), &metrics, http.MethodGet, http.NoBody, "application/json", ScrapeTimeout)
-	if err != nil {
-		return nil, err
-	}
-	return metrics, nil
-}
-
 func (p *pktvisorBackend) GetCapabilities() (map[string]interface{}, error) {
 	var taps interface{}
 	err := p.request("taps", &taps, http.MethodGet, http.NoBody, "application/json", TapsTimeout)
@@ -555,53 +289,10 @@ func (p *pktvisorBackend) GetCapabilities() (map[string]interface{}, error) {
 	return jsonBody, nil
 }
 
-func Register() bool {
-	backend.Register("pktvisor", &pktvisorBackend{
-		adminAPIProtocol: "http",
-	})
-	return true
-}
-
-func (p *pktvisorBackend) createOtlpMqttExporter(ctx context.Context) (component.MetricsExporter, error) {
-
-	if p.mqttClient != nil {
-		cfg := otlpmqttexporter.CreateConfigClient(p.mqttClient, p.otlpMetricsTopic, p.pktvisorVersion)
-		set := otlpmqttexporter.CreateDefaultSettings(p.logger)
-		// Create the OTLP metrics exporter that'll receive and verify the metrics produced.
-		exporter, err := otlpmqttexporter.CreateMetricsExporter(ctx, set, cfg)
-		if err != nil {
-			return nil, err
-		}
-		return exporter, nil
-	} else {
-		cfg := otlpmqttexporter.CreateConfig(p.mqttConfig.Address, p.mqttConfig.Id, p.mqttConfig.Key,
-			p.mqttConfig.ChannelID, p.pktvisorVersion, p.otlpMetricsTopic)
-		set := otlpmqttexporter.CreateDefaultSettings(p.logger)
-		// Create the OTLP metrics exporter that'll receive and verify the metrics produced.
-		exporter, err := otlpmqttexporter.CreateMetricsExporter(ctx, set, cfg)
-		if err != nil {
-			return nil, err
-		}
-		return exporter, nil
-	}
-
-}
-
-func createReceiver(ctx context.Context, exporter component.MetricsExporter, logger *zap.Logger) (component.MetricsReceiver, error) {
-	set := pktvisorreceiver.CreateDefaultSettings(logger)
-	cfg := pktvisorreceiver.CreateDefaultConfig()
-	// Create the Prometheus receiver and pass in the previously created Prometheus exporter.
-	receiver, err := pktvisorreceiver.CreateMetricsReceiver(ctx, set, cfg, exporter)
-	if err != nil {
-		return nil, err
-	}
-	return receiver, nil
-}
-
 func (p *pktvisorBackend) FullReset(ctx context.Context) error {
 
 	// force a stop, which stops scrape as well. if proc is dead, it no ops.
-	if state, _, _ := p.GetState(); state == backend.Running {
+	if state, _, _ := p.getProcRunningStatus(); state == backend.Running {
 		if err := p.Stop(ctx); err != nil {
 			p.logger.Error("failed to stop backend on restart procedure", zap.Error(err))
 			return err
@@ -616,4 +307,11 @@ func (p *pktvisorBackend) FullReset(ctx context.Context) error {
 	}
 
 	return nil
+}
+
+func Register() bool {
+	backend.Register("pktvisor", &pktvisorBackend{
+		adminAPIProtocol: "http",
+	})
+	return true
 }
